test: bind a deterministic geocoder globally to fix flaky "nearby" tests
The real Geocoder makes a live HTTP call to the Google Maps Geocoding API,
which is rate-limited and intermittently errors (~20% on the GCP dashboard).
Any test that geocodes a location was therefore at the mercy of the network —
most visibly Dashboard\BasicTest::testUpcomingEvents, which intermittently
failed with `Undefined array key "nearby"`: when a live geocode failed, the
event lost its coordinates and dropped out of the inner-joined nearby query,
so the `nearby` flag was never set.

Bind a deterministic GeocoderMock globally in TestCase::setUp() so no test ever
hits the live API, and remove the per-test GeocoderMock duplication in
CreateEventTest / OnlineEventsTest. Tests that exercise the geocode-failure
path (GroupEditGeocodeFallbackTest) keep their own failing mock, which
overrides the global binding. This also stops the test suite from contributing
to live Geocoding API traffic.

Verified green: BasicTest, CreateEventTest, OnlineEventsTest,
GroupEditGeocodeFallbackTest, AttendanceTest, DeleteEventTest,
AddRemoveVolunteerTest.
