+++ tests/Feature/Events/OnlineEventsTest.php
@@ -5,7 +5,6 @@ namespace Tests\Feature;
 use App\Device;
 use App\EventsUsers;
 use App\Group;
-use App\Helpers\Geocoder;
 use App\Network;
 use App\Notifications\NotifyRestartersOfNewEvent;
 use App\Party;
@@ -19,15 +18,6 @@ use Tests\TestCase;
 
 class OnlineEventsTest extends TestCase
 {
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        $this->app->bind(Geocoder::class, function () {
-            return new GeocoderMock();
-        });
-    }
-
     /** @test */
     public function a_host_can_create_an_online_event(): void
     {
