+++ tests/TestCase.php
@@ -11,6 +11,7 @@ use App\EventsUsers;
 use App\Group;
 use App\GroupNetwork;
 use App\GroupTags;
+use App\Helpers\Geocoder;
 use App\Images;
 use App\Network;
 use App\Party;
@@ -35,6 +36,7 @@ use Osteel\OpenApi\Testing\ValidatorBuilder;
 use Osteel\OpenApi\Testing\Exceptions\ValidationException;
 use ReflectionFunction;
 use Illuminate\Events\Dispatcher;
+use Tests\Support\GeocoderMock;
 
 abstract class TestCase extends BaseTestCase
 {
@@ -114,6 +116,14 @@ abstract class TestCase extends BaseTestCase
 
         $this->withoutExceptionHandling();
         app('honeypot')->disable();
+
+        // Bind a deterministic geocoder so the suite never makes live calls to the
+        // Google Maps Geocoding API (rate-limited / intermittently failing, which
+        // made any geocoding test flaky). Tests that need the geocode-failure path
+        // bind their own failing mock in setUp(), overriding this.
+        $this->app->bind(Geocoder::class, function () {
+            return new GeocoderMock();
+        });
 
         // Create categories defensively - only if they don't already exist
         if (!Category::where('idcategories', 111)->exists()) {
